% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq_query.R
\name{bq_query}
\alias{bq_query}
\title{Run a SQL query in BigQuery}
\usage{
bq_query(con, sql)
}
\arguments{
\item{con}{The BigQuery connection object}

\item{sql}{The SQL statement that is to be executed}
}
\description{
Run a SQL query in BigQuery
}
\examples{
bq_query(con, sql = "select * from users limit 1000")
}
\keyword{bigquery}
