% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer_theme.R
\name{buffer_theme}
\alias{buffer_theme}
\title{A function to apply a Buffer theme to a ggplot2 plot.}
\usage{
buffer_theme()
}
\description{
This applies the Buffer theme to your plot.
}
\examples{
ggplot(mtcars, aes(x = weight)) + geom_histogram() + buffer_theme()
}
\keyword{ggplot,}
\keyword{plot,}
\keyword{theme}

