% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffplot.R
\name{buffplot}
\alias{buffplot}
\title{Create a buffer plot with default colorblind-friendly color palette}
\usage{
buffplot(...)
}
\arguments{
\item{...}{Other arguments passed to \code{theme_minimal}}
}
\description{
Create a buffer plot with default colorblind-friendly color palette
}
\details{
This is a wrapper around ggplot2::ggplot() that adds default colors
}
\examples{
\dontrun{
buffplot(mtcars, aes(x = wt, y = mpg, color = gear)) +
    geom_point() +
    labs(title = "A Lovely Plot",
         subtitle = "What can the subtitle tell us?")
}

}
