% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_look.R
\name{check_looker_credentials}
\alias{check_looker_credentials}
\title{Return data from a Look}
\usage{
check_looker_credentials()
}
\arguments{
\item{look_id}{The numeric id of a Look found in its URL.}
}
\value{
A dataframe containing data from a Look.
}
\description{
\code{get_look(look_id)} returns a PostgreSQL connection.
}
\examples{
df <- get_look(3161)
}
\keyword{looker}

