% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bigquery_table.R
\name{create_bigquery_table}
\alias{create_bigquery_table}
\title{Create a new table in BigQuery if it does not exist.}
\usage{
create_bigquery_table(project_id, dataset_name, table_name, df_to_upload)
}
\arguments{
\item{project_id}{The BigQuery project ID, e.g. 'buffer-data'}

\item{dataset_name}{The name of the dataset in which you wish to create the table, e.g. 'temp_google_sheets'}

\item{table_name}{The name of the table you wish to create in BigQuery.}

\item{df_to_upload}{A dataframe that is to be inserted into the new table in BigQuery.}
}
\description{
Create a new table in BigQuery if it does not exist.
}
\examples{
create_bigquery_table(project_id = "buffer-data",
                      dataset_name = "temp_google_sheets",
                      table_name = "ig_profile_deprecation_list_c1_free",
                      df_to_upload = list_c1_free)
}
\keyword{bigquery}
