% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_batch.R
\name{create_empty_table}
\alias{create_empty_table}
\title{Write a data frame into Redshift}
\usage{
create_empty_table(con, tn, df)
}
\arguments{
\item{con}{Redshift connection.}

\item{df}{Data frame to write.}

\item{table_name}{Name of Reshift table you want to write to.}
}
\value{
A new or updated RS table.
}
\description{
\code{insert_batch(con, table_name, df)} writes the data frame into Redshift.
}
\section{Warning}{

Do not operate heavy machinery within 8 hours of using this function.
}
\examples{
con <- redshift_connect()
insert_batch(con, "mrr_forecast", forecast_df);
}
\keyword{buffer,}
\keyword{redshift}
\keyword{write,}

