% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain_prop_test.R
\name{explain_prop_test}
\alias{explain_prop_test}
\title{Explain proportion test}
\usage{
explain_prop_test(prop_test)
}
\arguments{
\item{prop_test}{The proportion test object.}
}
\value{
A string containing the explanation of the results of the test.
}
\description{
\code{explain_prop_test(prop_test)} returns a string explaining the results of a two-sample proportion test.
}
\examples{
explain_prop_test(prop_test)
}
\keyword{proportion}
\keyword{test}
