% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_customer_mrr_events.R
\name{get_customer_mrr_events}
\alias{get_customer_mrr_events}
\title{Return customer activities from ChartMogul}
\usage{
get_customer_mrr_events(uuid)
}
\arguments{
\item{uuid}{The ChartMogul UUID of the customer.}
}
\value{
A dataframe containing data from ChartMogul.
}
\description{
\code{get_customer_mrr_events(uuid = "cus_123")} returns a data frame with custoemr activites
}
\examples{
activites <- get_customer_mrr_events("cus_123")
}
\keyword{chartmogul}
