% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forecast.R
\name{get_forecast}
\alias{get_forecast}
\title{Forecast time series data}
\usage{
get_forecast(df, days, seasonality)
}
\arguments{
\item{df}{A time series data frame with a date and value column.}

\item{days}{The number of days you want to forecast out to.}

\item{seasonality}{The seasonality frequency (eg 7 for weekly seasonality).}
}
\value{
A new dataframe containing the forecasts.
}
\description{
\code{get_forecast(df, days, seasonality)} returns a dataframe with the forecast means and bounds.
}
\examples{
forecasts <- get_forecast(df, days = 30, seasonality = 7)
}
\keyword{forecast,}
\keyword{series}
\keyword{time}

