% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_batch.R
\name{insert_batch}
\alias{insert_batch}
\title{Write a data frame into Redshift}
\usage{
insert_batch(con, table_name, df, size = 100L)
}
\arguments{
\item{con}{Redshift connection.}

\item{table_name}{Name of Reshift table you want to write to.}

\item{df}{Data frame to write.}
}
\value{
A new or updated RS table.
}
\description{
\code{insert_batch(con, table_name, df)} writes the data frame into Redshift.
}
\examples{
con <- redshift_connect()
insert_batch(con, "mrr_forecast", forecast_df);
}
\keyword{buffer,}
\keyword{redshift}
\keyword{write,}
