% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_db.R
\name{query_db}
\alias{query_db}
\title{A function to query Redshift}
\usage{
query_db(query, connection)
}
\arguments{
\item{BUFFER_REDSHIFT_USER}{Your Buffer redshift user.}

\item{BUFFER_REDSHIFT_PWD}{Your Buffer redshift password.}
}
\description{
This function connects you to buffermetrics
}
\examples{
conn <- redshift_connect()
query <- "select count(*) from users"
plans <- query_db(query=query, connection = conn)
}
\keyword{buffer,}
\keyword{connect,}
\keyword{query}
\keyword{redshift,}
