% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift_connect.R
\name{redshift_connect}
\alias{redshift_connect}
\title{Create a connection to Redshift}
\usage{
redshift_connect()
}
\arguments{
\item{BUFFER_REDSHIFT_USER}{Your Buffer redshift user.}

\item{BUFFER_REDSHIFT_PWD}{Your Buffer redshift password.}
}
\value{
The PostgreSQL connection.
}
\description{
\code{redshift_connect} returns a PostgreSQL connection.
}
\section{Warning}{

Do not operate heavy machinery within 8 hours of using this function.
}
\examples{
conn <- redshift_connect()
}
\keyword{buffer,}
\keyword{connect,}
\keyword{redshift}

