% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain_prop_test.R
\docType{data}
\name{stmt1}
\alias{stmt1}
\title{Minimal ggplot2 theme using the Roboto Condensed font}
\format{An object of class \code{character} of length 1.}
\usage{
stmt1
}
\arguments{
\item{prop_test}{the proportion test object.}
}
\description{
Minimal ggplot2 theme using the Roboto Condensed font
}
\details{
This function attempts to communicate the results of a proportion test in a clear manner.
}
\examples{
\dontrun{
ptest <- prop.test(n_success, n_trials)
explain_prop_test(ptest)
}

}
\keyword{datasets}
