% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_to_redshift.R
\name{write_to_redshift}
\alias{write_to_redshift}
\title{Replace a table in Redshift.}
\usage{
write_to_redshift(df, table_name, bucket_name, option = "replace")
}
\arguments{
\item{df}{The R dataframe you want to write to Redshift}

\item{table_name}{The name of the Redshift table.}

\item{bucket_name}{The name of the s3 bucket.}

\item{option}{Determine whehter you want to replace or append table.}
}
\value{
Doesn't return anything. Writes data to Redshift
}
\description{
\code{redshift_connect} returns a PostgreSQL connection.
}
\section{Warning}{

Do not operate heavy machinery within 8 hours of using this function.
}

\examples{
write_to_redshift(my_data, "redshift_table_name", "new_bucket_name", "replace")
}
\keyword{buffer,}
\keyword{connect,}
\keyword{redshift,}
\keyword{write.}
