% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer_theme.R
\name{buffer_theme}
\alias{buffer_theme}
\title{Minimal ggplot2 theme using the Roboto Condensed font}
\usage{
buffer_theme()
}
\arguments{
\item{base_size}{base font size}

\item{strip_text_size, strip_text_margin}{plot strip text size and margin}

\item{subtitle_size, subtitle_margin}{plot subtitle size and margin}

\item{plot_title_size, plot_title_margin}{plot title size and margin}

\item{...}{Other arguments passed to \code{theme_minimal}}
}
\description{
Minimal ggplot2 theme using the Roboto Condensed font
}
\details{
The Roboto Condensed and Roboto Bold fonts are both Google fonts;
they can be found at \url{https://fonts.google.com/specimen/Roboto+Condensed}
and \url{https://fonts.google.com/specimen/Roboto}. These fonts must be
installed locally on your computer for this theme to work.
}
\examples{
\dontrun{
library(ggplot2)

ggplot(mtcars, aes(wt, mpg)) +
    geom_point() +
    labs(title = "A Lovely Plot",
         subtitle = "What can the subtitle tell us?") +
    buffer_theme()
}

}
