% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_look.R
\name{get_look}
\alias{get_look}
\title{Return data from a Look}
\usage{
get_look(look_id, id = Sys.getenv("LOOKER_API3_CLIENT_ID"),
  secret = Sys.getenv("LOOKER_API3_CLIENT_SECRET"))
}
\arguments{
\item{look_id}{The numeric id of a Look found in its URL.}
}
\value{
A dataframe containing data from a Look.
}
\description{
\code{get_look(look_id)} returns a PostgreSQL connection.
}
\examples{
client_id <- Sys.getenv("LOOKER_API3_CLIENT_ID")
client_secret <- Sys.getenv("LOOKER_API3_CLIENT_SECRET")
df <- get_look(3990, client_id, client_secret)
}
\keyword{looker}
