% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mrr_metrics.R
\name{get_mrr_metrics}
\alias{get_mrr_metrics}
\title{Return data from a Look}
\usage{
get_mrr_metrics(metric = "all", start_date, end_date, interval)
}
\arguments{
\item{metric}{The endpoint that specifies the type of metric, e.g. all, mrr, customers, ltv, customer-churn-rate, mrr-churn-rate}

\item{start_date}{The start date of the required period of data. An ISO-8601 formatted date, e.g. 2015-05-12.}

\item{end_date}{The end date of the required period of data. An ISO-8601 formatted date, e.g. 2015-05-12.
#' @param interval One of day, week, or month (default).}
}
\value{
A dataframe containing data from ChartMogul.
}
\description{
\code{get_mrr_metrics(metric = "all", start_date = "2019-01-01", end_date = "2019-03-01")} returns a data frame with results.
}
\examples{
df <- get_mrr_metrics(metric = "all", start_date = "2019-01-01", end_date = "2019-03-01", interval = "day")
}
\keyword{chartmogul}
