% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_names.R
\name{safe_names}
\alias{safe_names}
\title{Return lower case formatted column names}
\usage{
safe_names(names)
}
\arguments{
\item{names}{The column names that you wish to transform.}
}
\value{
A list of nicely formatted names.
}
\description{
\code{colnames(df) <- safe_names(colnames(df))} sets columns names to nice format.
}
\examples{
colnames(df) <- safe_names(colnames(df))
}
\keyword{columns,}
\keyword{format,}
\keyword{safe}

